+++ packages/asciitorium/src/examples/CelticBorderExample.tsx
@@ -2,7 +2,7 @@ import { Text, Component, Column, CelticBorder } from '../index';
 
 export const CelticBorderExample = () => {
   return (
-    <Component layout="aligned" label="CelticBorder Example:" border>
+    <Component height="fill" layout="aligned" label="CelticBorder Example:" border>
         <CelticBorder align="top-left" />
         <CelticBorder align="bottom-left" />
         <CelticBorder align="top-right" />